gene_splicer/requirements.txt
git+https://github.com/cfe-lab/MiCall.git@v7.7.0#egg=gotoh&subdirectory=micall/alignment
git+https://github.com/cfe-lab/iva.git@v1.1.1
numpy==1.18.4
python-Levenshtein==0.12.0
pandas==1.0.5
matplotlib==3.3.3
requests==2.24.0
pyyaml==5.3.1